Making recipe_name process removed from recipes.management
