+++ .ebextensions/01_command_files.config
@@ -44,5 +44,5 @@ files:
       if [ -f /tmp/create_ingredient ]
       then
         rm /tmp/create_ingredient
-        sudo docker exec `sudo docker ps -q` /srv/project/app/manage.py create_ingredient_ko
+        sudo docker exec `sudo docker ps -q` /srv/project/app/manage.py create_ingredient
       fi
