+++ app/recipes/management/commands/__init__.py
@@ -1,0 +1,2 @@
+from .create_ingredient import *
+from .create_ingredient_eng import *
